+++ src/Chain.h
@@ -40,7 +40,7 @@ class Chain
       return (this->isRT_Chain && !(right.isRT_Chain));
     }
     
-    uint8_t mu_u_at_lvl (uint8_t lvl); // returns the amount of cpu required for placing this chain at level lvl
+    uint16_t mu_u_at_lvl (uint8_t lvl); // returns the amount of cpu required for placing this chain at level lvl
 };
 
 /* 
@@ -52,7 +52,7 @@ bool findChainInSet (set<Chain> setOfChains, int32_t id, Chain& foundChain);
 class RT_Chain : public Chain
 {
 public:
-  static const vector<uint8_t> mu_u; // mu_u[i] will hold the # of cpu units required for placing an RT chain on a DC in level i
+  static const vector<uint16_t> mu_u; // mu_u[i] will hold the # of cpu units required for placing an RT chain on a DC in level i
   static const uint8_t mu_u_len;
 	static const vector<uint16_t> cpuCostAtLvl; // cpuCostAtLvl[i] will hold the cost of placing an RT chain on a DC in level i
   RT_Chain (int32_t id, vector <int16_t> S_u);
@@ -61,7 +61,7 @@ public:
 class Non_RT_Chain: public Chain
 {
   public:
-	  static const vector<uint8_t> mu_u; // mu_u[i] will hold the # of cpu units required for placing an RT chain on a DC in level i
+	  static const vector<uint16_t> mu_u; // mu_u[i] will hold the # of cpu units required for placing an RT chain on a DC in level i
 	  static const uint8_t  mu_u_len;
 		static const vector<uint16_t> cpuCostAtLvl; // cpuCostAtLvl[i] will hold the cost of placing a non- RT chain on a DC in level i
     Non_RT_Chain (int32_t id, vector <int16_t> S_u);
