+++ src/MyConfig.cc
@@ -32,7 +32,7 @@ void MyConfig::printToLog (string str)
 	logFile << str;
 }
 
-vector<uint16_t> MyConfig::scalarProdcut (const vector<uint8_t> &vec1, const vector<uint16_t> &vec2) 
+vector<uint16_t> MyConfig::scalarProdcut (const vector<uint16_t> &vec1, const vector<uint16_t> &vec2) 
 { 
 	vector <uint16_t> res;
 	for (uint8_t i(0); i<vec1.size(); i++) {
